+++ src/providers/clicky/clicky.js
@@ -4,7 +4,9 @@
 
 analytics.addProvider('Clicky', {
 
-    settings : {},
+    settings : {
+        siteId : null
+    },
 
 
     // Initialize
@@ -24,6 +26,14 @@ analytics.addProvider('Clicky', {
             s.src = '//static.getclicky.com/js';
             (document.getElementsByTagName('head')[0] || document.getElementsByTagName('body')[0]).appendChild(s);
         })();
+    },
+
+
+    // Track
+    // -----
+
+    track : function (event, properties) {
+        if (window.clicky) window.clicky.log(window.location.href, event);
     }
 
 });